Added Player List Functionality.
Added functionality to the player list.

When a player joins a game room their name will be added to the player list of every currently joined player. When a player leaves their name will be removed from every other players list.

The name of the client is always the first name listed.